Makefile: Add an "install" target
Add an "install" target that installs the program and license file to a
typical prefix, defaulting to /usr/local but overridable with standard
variables.

The use of "mkdir -p" here is not portable but works on macOS which is
the only OS this Makefile is expected to be used with.