Makefile improvements (#84)
* Makefile: Use $CC instead of gcc

Use $CC instead of gcc. There isn't a real gcc on macOS; it's a link to
clang. The user may wish to use a different compiler specified in $CC.
If not, make defaults to cc, which is also a link to clang.

* Makefile: Use $@ and $<

Use $@ for the target and $< for the first prerequisite rather than
repeating that information in the rule.

* Makefile: Use $CFLAGS, $CPPFLAGS, and $LDFLAGS

Use $CFLAGS, $CPPFLAGS, and $LDFLAGS that the user may have specified.
Add -I. first to ensure that the local directory is searched for include
files first, before any -I paths that might be specified in $CFLAGS or
$CPPFLAGS.

* Makefile: Add an "all" target

Add an "all" target. Users familiar with automake may expect to be able
to run "make all".

* Makefile: Add a "clean" target

Add a "clean" target that removes what was built.

* Makefile: Add an "install" target

Add an "install" target that installs the program and license file to a
typical prefix, defaulting to /usr/local but overridable with standard
variables.

The use of "mkdir -p" here is not portable but works on macOS which is
the only OS this Makefile is expected to be used with.

* Add debug target, remove some boilerplate code that does not apply to this project

Co-authored-by: Jake Hilborn <jakehilborn@gmail.com>