Update README to document how to compile and install the latest code
NOTE: This relies on https://github.com/jakehilborn/homebrew-jakehilborn/pull/1 being merged.

The current github release is very out of date and from 2019. It's missing a few bug fixes. The inaccurate support for "scaling:off" comes to mind.

But instead of complaining about the lack of a recent releases, we can instead just use homebrew to make it easy for users to compile and install the latest version of the code currently on github. This is particularly simple for displayplacer since it has no special build dependencies.

Considering the pace of commits and relatively stable nature of master, I don't think this introduces too much risk.  (I think `install --head` is an under utilized feature of homebrew taps.  It obviates much of the need for the maintainer to constantly update the github releases!)

This commit just updates the instructions to tell users to install with `brew install --head displayplacer`. It of course relies on the PR mentioned above also being merged.

Also note that users that already tapped the homebrew tap will need to update it for `install --head` to work.  This should happen automatically after a `brew update`, which is something homebrew also does automatically after a certain amount of time.